Add loading indicator to the test liveview
This PR adds a loading animation indicating that the live view is being
loaded. Currently only a blank box is displayed, visible in the webui
before liveview content is received by the responsible canvas element.

Test cases are also added to confirm the elements are displayed and
hidden as expected.